+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="2.6.3"></a>
+
+# [2.6.3](https://gitlab.com/atjdev/ezio-api/-/merge_requests/142) (2022-04-28)
+
+### Features
+
+- [**M014**] **warping**: add some queries for AT4 form.
+- add ifs shop order operation entity and resolver.
+
 <a name="2.6.2"></a>
 
 # [2.6.2](https://gitlab.com/atjdev/ezio-api/-/merge_requests/140) (2022-04-27)